requirements.txt
# Core Dependencies
streamlit
python-dotenv
pyyaml
pydantic

# LLM Providers
openai
anthropic
cohere
langchain
langchain-community
langchain-openai
langgraph
langchain-core
langchain-text-splitters

# Vector Databases
chromadb
# faiss-cpu
# pinecone-client

# Document Processing
pypdf
python-docx
tiktoken

# Embeddings (minimal)
sentence-transformers
transformers

# Evaluation (optional - comment out if not needed)
# ragas
# datasets

# Monitoring & Logging
loguru

# Utilities
numpy
pandas
requests
tavily-python

# AWS Deployment (optional - comment out if not needed)
# boto3
# sagemaker

# Azure Deployment
azure-storage-blob
azure-identity
# azure-mgmt-containerinstance
# azure-mgmt-web

# HuggingFace Deployment (optional)
huggingface-hub
